# Fix slow IsCheckpointImage making podman run slow