@rhatdan 2022-10-06
I think this is correct, although I am worried about podman-remote versus podman.  Since I think this is only getting the OCI Runtime from the client.  I am not sure if this runtime is sent to the server side. Calling info for this information does seem to be way overkill though.

@mheon PTAL