@rhatdan 2022-10-06
I now believe this function should be called on the server side not on the client side anyways.